requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: win-64
aiofiles=0.8.0=pypi_0
anyio=3.7.1=pypi_0
argon2-cffi=21.3.0=pyhd3eb1b0_0
argon2-cffi-bindings=21.2.0=py310h2bbff1b_0
asgiref=3.7.2=pypi_0
asttokens=2.2.1=pyhd8ed1ab_0
attrs=22.1.0=py310haa95532_0
backcall=0.2.0=pyh9f0ad1d_0
backports=1.0=pyhd8ed1ab_3
backports.functools_lru_cache=1.6.5=pyhd8ed1ab_0
beautifulsoup4=4.12.2=py310haa95532_0
blas=1.0=mkl
bleach=4.1.0=pyhd3eb1b0_0
bottleneck=1.3.5=py310h9128911_0
brotli=1.0.9=h2bbff1b_7
brotli-bin=1.0.9=h2bbff1b_7
bzip2=1.0.8=he774522_0
ca-certificates=2023.05.30=haa95532_0
carvekit=4.1.0=pypi_0
certifi=2023.7.22=pypi_0
cffi=1.15.1=py310h2bbff1b_3
charset-normalizer=2.0.12=pypi_0
click=8.1.6=pypi_0
colorama=0.4.6=pyhd8ed1ab_0
comm=0.1.4=pyhd8ed1ab_0
contourpy=1.0.5=py310h59b6b97_0
cycler=0.11.0=pyhd3eb1b0_0
debugpy=1.6.7=py310hd77b12b_0
decorator=5.1.1=pyhd8ed1ab_0
defusedxml=0.7.1=pyhd3eb1b0_0
entrypoints=0.4=py310haa95532_0
exceptiongroup=1.1.3=pypi_0
executing=1.2.0=pyhd8ed1ab_0
fastapi=0.78.0=pypi_0
fonttools=4.25.0=pyhd3eb1b0_0
freetype=2.12.1=ha860e81_0
giflib=5.2.1=h8cc25b3_3
h11=0.14.0=pypi_0
icc_rt=2022.1.0=h6049295_2
idna=3.4=py310haa95532_0
importlib-metadata=6.8.0=pyha770c72_0
importlib_metadata=6.8.0=hd8ed1ab_0
intel-openmp=2023.1.0=h59b6b97_46319
ipykernel=6.25.1=pyh6817e22_0
ipympl=0.9.3=pypi_0
ipython=8.14.0=pyh08f2357_0
ipython_genutils=0.2.0=pyhd3eb1b0_1
ipywidgets=8.1.0=pypi_0
jedi=0.19.0=pyhd8ed1ab_0
jinja2=3.1.2=py310haa95532_0
joblib=1.2.0=py310haa95532_0
jpeg=9e=h2bbff1b_1
jsonschema=4.17.3=py310haa95532_0
jupyter_client=7.4.9=py310haa95532_0
jupyter_core=5.3.1=py310h5588dad_0
jupyter_server=1.23.4=py310haa95532_0
jupyterlab-widgets=3.0.8=pypi_0
jupyterlab_pygments=0.1.2=py_0
kiwisolver=1.4.4=py310hd77b12b_0
lerc=3.0=hd77b12b_0
libbrotlicommon=1.0.9=h2bbff1b_7
libbrotlidec=1.0.9=h2bbff1b_7
libbrotlienc=1.0.9=h2bbff1b_7
libdeflate=1.17=h2bbff1b_0
libffi=3.4.4=hd77b12b_0
libiconv=1.16=h2bbff1b_2
libpng=1.6.39=h8cc25b3_0
libsodium=1.0.18=h8d14728_1
libtiff=4.5.0=h6c2663c_2
libwebp=1.2.4=hbc33d0d_1
libwebp-base=1.2.4=h2bbff1b_1
libxml2=2.10.3=h0ad7f3c_0
libxslt=1.1.37=h2bbff1b_0
loguru=0.6.0=pypi_0
lxml=4.9.2=py310h2bbff1b_0
lz4-c=1.9.4=h2bbff1b_0
markupsafe=2.1.1=py310h2bbff1b_0
matplotlib-base=3.7.1=py310h4ed8f06_1
matplotlib-inline=0.1.6=pyhd8ed1ab_0
mistune=0.8.4=py310h2bbff1b_1000
mkl=2023.1.0=h6b88ed4_46357
mkl-service=2.4.0=py310h2bbff1b_1
mkl_fft=1.3.6=py310h4ed8f06_1
mkl_random=1.2.2=py310h4ed8f06_1
munkres=1.1.4=py_0
nbclassic=0.5.5=py310haa95532_0
nbclient=0.5.13=py310haa95532_0
nbconvert=6.5.4=py310haa95532_0
nbformat=5.7.0=py310haa95532_0
nest-asyncio=1.5.6=pyhd8ed1ab_0
notebook=6.5.4=py310haa95532_1
notebook-shim=0.2.2=py310haa95532_0
numexpr=2.8.4=py310h2cd9be0_1
numpy=1.22.4=pypi_0
numpy-base=1.25.2=py310h65a83cf_0
opencv-python=4.5.5.64=pypi_0
openssl=3.0.10=h2bbff1b_0
packaging=23.1=pyhd8ed1ab_0
pandas=1.5.3=py310h4ed8f06_0
pandocfilters=1.5.0=pyhd3eb1b0_0
parso=0.8.3=pyhd8ed1ab_0
pickleshare=0.7.5=py_1003
pillow=9.0.1=pypi_0
pip=23.2.1=py310haa95532_0
platformdirs=3.10.0=pyhd8ed1ab_0
plotly=5.16.0=py_0
prometheus_client=0.14.1=py310haa95532_0
prompt-toolkit=3.0.39=pyha770c72_0
prompt_toolkit=3.0.39=hd8ed1ab_0
psutil=5.9.0=py310h2bbff1b_0
pupil-apriltags=1.0.4.post10=pypi_0
pure_eval=0.2.2=pyhd8ed1ab_0
pycparser=2.21=pyhd3eb1b0_0
pydantic=1.9.2=pypi_0
pygments=2.16.1=pyhd8ed1ab_0
pyparsing=3.0.9=py310haa95532_0
pyrsistent=0.18.0=py310h2bbff1b_0
python=3.10.12=he1021f5_0
python-dateutil=2.8.2=pyhd8ed1ab_0
python-fastjsonschema=2.16.2=py310haa95532_0
python-multipart=0.0.6=pypi_0
python_abi=3.10=2_cp310
pytz=2022.7=py310haa95532_0
pywin32=305=py310h2bbff1b_0
pywinpty=2.0.10=py310h5da7b33_0
pyzmq=23.2.1=py310h73ada01_0
requests=2.27.1=pypi_0
scikit-learn=1.3.0=py310h4ed8f06_0
scipy=1.11.1=py310h309d312_0
seaborn=0.12.2=py310haa95532_0
send2trash=1.8.0=pyhd3eb1b0_1
setuptools=62.3.4=pypi_0
six=1.16.0=pyh6c4a22f_0
sniffio=1.3.0=pypi_0
soupsieve=2.4=py310haa95532_0
sqlite=3.41.2=h2bbff1b_0
stack_data=0.6.2=pyhd8ed1ab_0
starlette=0.19.1=pypi_0
tbb=2021.8.0=h59b6b97_0
tenacity=8.2.2=py310haa95532_0
terminado=0.17.1=py310haa95532_0
threadpoolctl=2.2.0=pyh0d69192_0
tinycss2=1.2.1=py310haa95532_0
tk=8.6.12=h2bbff1b_0
torch=1.11.0+cu113=pypi_0
torchvision=0.12.0+cu113=pypi_0
tornado=6.2=py310he2412df_0
tqdm=4.64.1=pypi_0
traitlets=5.9.0=pyhd8ed1ab_0
typing=3.7.4.3=pypi_0
typing-extensions=4.7.1=hd8ed1ab_0
typing_extensions=4.7.1=pyha770c72_0
tzdata=2023c=h04d1e81_0
urllib3=1.26.16=pypi_0
uvicorn=0.17.6=pypi_0
vc=14.2=h21ff451_1
vs2015_runtime=14.27.29016=h5e58377_2
wcwidth=0.2.6=pyhd8ed1ab_0
webencodings=0.5.1=py310haa95532_1
websocket-client=0.58.0=py310haa95532_4
wheel=0.38.4=py310haa95532_0
widgetsnbextension=4.0.8=pypi_0
win32-setctime=1.1.0=pypi_0
winpty=0.4.3=4
xz=5.4.2=h8cc25b3_0
zeromq=4.3.4=h0e60522_1
zipp=3.16.2=pyhd8ed1ab_0
zlib=1.2.13=h8cc25b3_0
zstd=1.5.5=hd43e919_0
